+++ NEWS.md
@@ -2,6 +2,7 @@
 
 Bug fixes
 
+* `report.brmsfit()`: fix issue where report text was printed multiple times when different parameters had different priors (#543)
 * Fixed duplicated text output in `report()` for glmmTMB objects by addressing both regex pattern and redundant CI information concatenation in `report_info.lm()` (#481)
 * Fixed issue with missing effect size for the Intercept term in type 3 anova tables (#451)